pyproject.toml
[project]
name = "chess-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "fastphase>=2.0.4",
    "gymnasium>=1.2.3",
    "modal>=1.3.0.post1",
    "numpy>=2.4.0",
    "pydantic>=2.12.5",
    "python-chess>=1.999",
    "torch>=2.9.1",
    "uvicorn>=0.40.0",
    "verifiers>=0.1.8.post2",
    "wandb>=0.23.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
